Moves remaining static methods into helper class 🚛
…as requested by @sabieber during review.

OX-9272
