+++ src/main/java/sirius/biz/charts/BaseChart.java
@@ -8,7 +8,6 @@
 
 package sirius.biz.charts;
 
-import org.apache.batik.anim.dom.SVGDOMImplementation;
 import org.w3c.dom.Element;
 
 import java.awt.Dimension;
@@ -153,32 +152,4 @@ public abstract class BaseChart {
      * @return the SVG representation of the chart
      */
     public abstract Element toSvg(Dimension bounds);
-
-    /**
-     * Creates an empty SVG element with the relevant attributes set.
-     *
-     * @return an empty SVG element
-     */
-    protected static Element createSvgElement() {
-        return SVGDOMImplementation.getDOMImplementation()
-                                   .createDocument(SVGDOMImplementation.SVG_NAMESPACE_URI, TAG_SVG, null)
-                                   .getDocumentElement();
-    }
-
-    /**
-     * Creates an empty SVG element with the view box centered.
-     *
-     * @param bounds the dimensions of the viewport
-     * @return an empty SVG element with the view box centered
-     */
-    protected static Element createSvgElementWithCenteredViewbox(Dimension bounds) {
-        Element svgElement = createSvgElement();
-        svgElement.setAttribute(ATTRIBUTE_VIEW_BOX,
-                                String.format("%f %f %f %f",
-                                              -0.5 * bounds.width,
-                                              -0.5 * bounds.height,
-                                              (double) bounds.width,
-                                              (double) bounds.height));
-        return svgElement;
-    }
 }
