+++ src/main/java/sirius/biz/charts/PieChart.java
@@ -61,7 +61,7 @@ public class PieChart<N extends Number> extends BaseChart {
 
     @Override
     public Element toSvg(Dimension bounds) {
-        Element svgElement = createSvgElementWithCenteredViewbox(bounds);
+        Element svgElement = Charts.createSvgElementWithCenteredViewbox(bounds);
 
         Element pieGroupElement =
                 svgElement.getOwnerDocument().createElementNS(SVGDOMImplementation.SVG_NAMESPACE_URI, TAG_G);
